[refactor][update] flatten the `trap::hypervisor_supervisor` module
